--- src/components/Mirror.jsx
@@ -1,12 +1,0 @@
-import React from 'react';
-
-export default function Mirror ({count}) {
-    return (
-        <span>
-            <b>
-                Mirror:
-            </b>
-            {count}
-        </span>
-    );
-}
